+++ lib/strategy-factory.ts
@@ -1,4 +1,4 @@
-import { DbDialect, ShadtsConfig } from "./types";
+import { DbDialect, ShadrizzConfig } from "./types";
 import { BetterSqlite3PackageStrategy } from "../db-packages/better-sqlite3-package-strategy";
 import { Mysql2PackageStrategy } from "../db-packages/mysql2-package-strategy";
 import { PgPackageStrategy } from "../db-packages/pg-package-strategy";
@@ -6,7 +6,7 @@ import { sqliteDialectStrategy } from "../db-dialects/sqlite-dialect-strategy";
 import { mysqlDialectStrategy } from "../db-dialects/mysql-dialect-strategy";
 import { postgresqlDialectStrategy } from "../db-dialects/postgresql-dialect-strategy";
 
-export function packageStrategyFactory(opts: ShadtsConfig) {
+export function packageStrategyFactory(opts: ShadrizzConfig) {
   switch (opts.dbPackage) {
     case "better-sqlite3":
       return new BetterSqlite3PackageStrategy(opts);