+++ scripts/test-sqlite.sh
@@ -1,25 +1,25 @@
-SHADTS_PATH="$HOME/code/shadts/index.ts"
+SHADRIZZ_PATH="$HOME/code/shadrizz/index.ts"
 
-shadts() {
-    tsx "$SHADTS_PATH" "$@"
+shadrizz() {
+    tsx "$SHADRIZZ_PATH" "$@"
 }
 
 rm -rf ~/code/demo-sqlite
 cd ~/code
-shadts new demo-sqlite -p bun --latest
+shadrizz new demo-sqlite -p bun --latest
 cd ~/code/demo-sqlite
-shadts init -p bun --latest --db-dialect sqlite -pk cuid2 --auth-solution authjs --auth-providers github,google,postmark,nodemailer,credentials --admin
-cp ~/code/shadts-env/.env.local.sqlite .env.local
-shadts add tiptap
-shadts scaffold -a admin admin_scaffold -c integer_type:integer real_type:real text_type:text boolean_type:boolean bigint_type:bigint timestamp_type:timestamp file_type:file
-shadts scaffold -a private private_scaffold -c text_field:text integer_field:integer real_field:real boolean_field:boolean file_field:file timestamp_field:timestamp
-shadts scaffold -a public public_scaffold -c text_field:text integer_field:integer real_field:real boolean_field:boolean file_field:file timestamp_field:timestamp
-shadts scaffold -a admin category -c title:text
-shadts scaffold -a admin post_status -c status:text
-shadts scaffold -a admin post -c category:references_combobox post_status:references_select title:text likes:integer published_at:timestamp content:text_tiptap
-shadts scaffold -a admin tags -c name:text
-shadts scaffold -a admin posts_tags -c post:references tag:references
-shadts join -a admin posts posts_tags tags
+shadrizz init -p bun --latest --db-dialect sqlite -pk cuid2 --auth-solution authjs --auth-providers github,google,postmark,nodemailer,credentials --admin
+cp ~/code/shadrizz-env/.env.local.sqlite .env.local
+shadrizz add tiptap
+shadrizz scaffold -a admin admin_scaffold -c integer_type:integer real_type:real text_type:text boolean_type:boolean bigint_type:bigint timestamp_type:timestamp file_type:file
+shadrizz scaffold -a private private_scaffold -c text_field:text integer_field:integer real_field:real boolean_field:boolean file_field:file timestamp_field:timestamp
+shadrizz scaffold -a public public_scaffold -c text_field:text integer_field:integer real_field:real boolean_field:boolean file_field:file timestamp_field:timestamp
+shadrizz scaffold -a admin category -c title:text
+shadrizz scaffold -a admin post_status -c status:text
+shadrizz scaffold -a admin post -c category:references_combobox post_status:references_select title:text likes:integer published_at:timestamp content:text_tiptap
+shadrizz scaffold -a admin tags -c name:text
+shadrizz scaffold -a admin posts_tags -c post:references tag:references
+shadrizz join -a admin posts posts_tags tags
 npm run generate
 npm run migrate
 npx tsx scripts/create-user.ts user@example.com pw