+++ test/logs.test.ts
@@ -4,10 +4,10 @@ import { AdminProcessor } from "../processors/admin-processor";
 import { StripeProcessor } from "../processors/stripe-processor";
 import { BetterSqlite3PackageStrategy } from "../db-packages/better-sqlite3-package-strategy";
 import { PgPackageStrategy } from "../db-packages/pg-package-strategy";
-import { ShadtsConfig } from "../lib/types";
+import { ShadrizzConfig } from "../lib/types";
 import { Mysql2PackageStrategy } from "../db-packages/mysql2-package-strategy";
 
-const shadtsConfig: ShadtsConfig = {
+const shadrizzConfig: ShadrizzConfig = {
   authProviders: ["google", "github", "credentials", "nodemailer", "postmark"],
   install: false,
   packageManager: "npm",
@@ -22,31 +22,31 @@ const shadtsConfig: ShadtsConfig = {
 };
 
 test("better sqlite3 package strategy", () => {
-  const s = new BetterSqlite3PackageStrategy(shadtsConfig);
+  const s = new BetterSqlite3PackageStrategy(shadrizzConfig);
   s.printCompletionMessage();
 });
 
 test("pg package strategy", () => {
-  const s = new PgPackageStrategy(shadtsConfig);
+  const s = new PgPackageStrategy(shadrizzConfig);
   s.printCompletionMessage();
 });
 
 test("mysql2 package strategy", () => {
-  const s = new Mysql2PackageStrategy(shadtsConfig);
+  const s = new Mysql2PackageStrategy(shadrizzConfig);
   s.printCompletionMessage();
 });
 
 test("auth processor", () => {
-  const a = new AuthProcessor(shadtsConfig);
+  const a = new AuthProcessor(shadrizzConfig);
   a.printCompletionMessage();
 });
 
 test("admin processor", () => {
-  const a = new AdminProcessor(shadtsConfig);
+  const a = new AdminProcessor(shadrizzConfig);
   a.printCompletionMessage();
 });
 
 test("stripe processor", () => {
-  const p = new StripeProcessor(shadtsConfig);
+  const p = new StripeProcessor(shadrizzConfig);
   p.printCompletionMessage();
 });